Add textures endpoint to fetch client jar
Mojang's new launcher manifest stuff made it impossible to fetch
a specific client jar with a single wget.

This commit adds a new endpoint to the overviewer flask app, called
`textures`, which HTTP redirects to the client.jar of a version of
one's choosing, or the latest release one with "latest".

E.g.:
 - /textures/latest -> 302 to latest client.jar
 - /textures/1.13.2 -> 302 to 1.13.2 client.jar

 The launcher manifest is cached for 20 minutes, the client manifests
 are cached for 1 hour.